Fix seeding issues
