+++ frontend/src/components/app.js
@@ -4,13 +4,11 @@ import { Switch } from "react-router-dom";
 import NavBarContainer from "./Nav/navbar_container";
 import Splash from "./Splash/splash";
 import Modal from "./Modal/modal";
-import ProductIndexByCategory from './Products/index_by_category_container';
+import ProductIndexByCategory from './Products/index_products_by_category'
 import "./app.css";
 
 import ProductIndexContainer from "./Products/product_index_container";
 import ProductUserContainer from "./Products/product_user_index_container";
-import LoginFormContainer from "./Session/login_form_container";
-import SignupFormContainer from "./Session/signup_form_container";
 
 import ProductShowContainer from './Product_Show/product_show_container'
 import "./app.css";
@@ -23,7 +21,7 @@ const App = () => (
 			<ProtectedRoute exact path="/products/:product_id" component={ProductShowContainer} />
 			<ProtectedRoute exact path="/users/:user_id/products" component={ProductUserContainer} />
       <ProtectedRoute exact path="/products" component={ProductIndexContainer} />
-       <ProtectedRoute 
+      <ProtectedRoute 
         exact 
         path="products/cat/:category" 
         component={ProductIndexByCategory}
